LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-1_4_1-expected.txt

setTimeout() and setInterval() should not run without 'unsafe-eval' script-src directive.


PASS window.setTimeout()
PASS window.setInterval()
FAIL Test that securitypolicyviolation event is fired assert_equals: expected "script-src" but got "script-src 'self' 'unsafe-inline'"
